arm64: dts: rockchip: fixes sd card on rk3399pro evb board
As someone feedback some sd-cards can't work on rk3399pro board, due to power
sequency or timing isn't suit for the spec.

Especially for the SD card 3.0 that probability can't be identified, the voltage
changes from 3.0v to 1.8v is too long, It caused by that the internal resistance
and capacitance of rk809 form RC circuit.

Change-Id: I4a4f6153c39fe022fd938b0567389be5beec5881
Signed-off-by: Caesar Wang <wxt@rock-chips.com>